fix: change data type to int64 in load_cluster_ids for improved compatibility

+++ IVF.py
@@ -132,9 +132,9 @@ def load_cluster_ids(filename, cluster_index, lengths_array, ids_offset):
 
         # Read that slice only
         f.seek(ids_offset + start * 4)
-        data = np.frombuffer(f.read(length * 4), dtype=np.uint32)
+        data = np.frombuffer(f.read(length * 4), dtype=np.int64)
 
-        return np.array(data, dtype=np.int64)
+        return np.array(data)
 
 
 def search(vec_db, query_vector, k=5, batch_size=500):
